docchat-backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain-community==0.3.25
langchain-chroma==0.2.4
langchainhub==0.1.21
langchain-huggingface==0.3.0
# llama-cpp-python==0.3.9
python-dotenv==1.1.0
together==1.5.13
# huggingface-hub
sentence-transformers
PyMuPDF
pdf2image 
Pillow